removendo logs

+++ Mining.py
@@ -43,7 +43,6 @@ class Mining:
 		self.pagination['current'] = 1
 		for file in files:
 			id = file.split('.xml')[0]
-			print(id)
 			self.saveItem(id)
 
 	def fetchIds(self, ids):
@@ -112,20 +111,20 @@ class Mining:
 				'Accept': 'application/json'
 			}
 
-			try:
-				print('PARSING...')
-				parser = Parser(f'/../{self.db}/xml/depressive/{id}.xml')
-				parser.parse()
-				data['publication'] = parser.dict
-			except Exception as e:
-				print(str(e))
-				os.remove(f'{self.path}/{id}.xml')
-				return
-			
-			if(data['publication']):
-				requests.post(url = f'{API}/publications', data = json.dumps(data), headers = headers) 
-
-			# File.cls()
+			res = requests.get(url = f'{API}/publications/{id}', headers = headers)
+			if(res.status_code != 200):
+				try:
+					parser = Parser(f'/{self.db}/xml/depressive/{id}.xml')
+					parser.parse()
+					data['publication'] = parser.dict
+				except Exception as e:
+					print(str(e))
+					os.remove(f'{self.path}/{id}.xml')
+					return
+				
+				if(data['publication']):
+					requests.post(url = f'{API}/publications', data = json.dumps(data), headers = headers) 
+			File.cls()
 			total = self.pagination['current']/self.pagination['count'] * 100
 
 			print('\t<<< MINING >>>')
